+++ src/app/core/store/app-layout/index.ts
@@ -1,2 +1,3 @@
 export * from './app-layout.reducer';
 export * from './app-layout.actions';
+export * from './app-layout.selectors';